Move variant specific code
