+++ variants/xiao_c6/XiaoC6Board.cpp
@@ -47,3 +47,5 @@ mesh::LocalIdentity radio_new_identity() {
   RadioNoiseListener rng(radio);
   return mesh::LocalIdentity(&rng);  // create new random identity
 }
+
+
