+++ variants/xiao_c6/platformio.ini
@@ -28,6 +28,7 @@ build_flags =
   -D DISABLE_WIFI_OTA=1
 build_src_filter = ${esp32c6_base.build_src_filter}
   +<../variants/xiao_c6>
+  +<XiaoC6Board.cpp>
 
 [env:Xiao_C6_Repeater]
 extends = Xiao_C6
